requirements.txt
absl-py==2.1.0
accelerate==1.1.1
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==2.4.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
-e git+https://github.com/spatank/compositional-rl-synth-data.git@f50b53d07afb94914df96e7ec196faf220d8bed3#egg=compositional_rl_synth_data
-e git+https://github.com/Lifelong-ML/CompoSuite.git@1fa36f67f31aeccc9ef75748bfc797960e044a86#egg=composuite
contourpy==1.3.0
cycler==0.12.1
Cython==0.29.37
d3rlpy==1.1.1
dataclasses-json==0.6.7
debugpy==1.8.8
decorator==5.1.1
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.25
docker-pycreds==0.4.0
einops==0.8.0
ema-pytorch==0.7.6
etils==1.5.2
exceptiongroup==1.2.2
executing==2.1.0
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.16.1
fonttools==4.55.1
fsspec==2024.10.0
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.8.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gym==0.25.2
gym-notices==0.0.8
gymnasium==1.0.0
h5py==3.12.1
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
labmaze==1.0.6
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
minari==0.5.1
mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
ml_dtypes==0.5.0
mpmath==1.3.0
mujoco==3.2.6
mujoco-py==2.1.2.14
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pynndescent==0.5.13
pynput==1.7.7
pyobjc-core==10.3.2
pyobjc-framework-ApplicationServices==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreText==10.3.2
pyobjc-framework-Quartz==10.3.2
PyOpenGL==3.1.7
pyparsing==3.2.0
pyrallis==0.3.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qpsolvers==4.4.0
quadprog==0.1.13
requests==2.32.3
rich==13.9.4
robosuite==1.4.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.19.0
setproctitle==1.3.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
structlog==24.4.0
sympy==1.13.1
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.5.1
torchdiffeq==0.2.5
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.13.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==1.26.20
wandb==0.18.7
wcwidth==0.2.13
widgetsnbextension==4.0.13
zipp==3.21.0
